move revenue to aggregates
