+++ ogusa/aggregates.py
@@ -9,6 +9,7 @@ Functions to compute economic aggregates.
 
 # Packages
 import numpy as np
+from . import tax
 
 '''
 ------------------------------------------------------------------------
@@ -197,3 +198,83 @@ def get_C(c, params):
     elif method == 'TPI':
         aggC = aggC_presum.sum(1).sum(1)
     return aggC
+
+
+def revenue(r, w, b, n, BQ, Y, L, K, factor, params):
+    '''
+    Gives lump sum transfer value.
+    Inputs:
+        r           = [T,] vector, interest rate
+        w           = [T,] vector, wage rate
+        b           = [T,S,J] array, wealth holdings
+        n           = [T,S,J] array, labor supply
+        BQ          = [T,J] array, bequest amounts
+        factor      = scalar, model income scaling factor
+        params      = length 12 tuple, (e, lambdas, omega, method, etr_params,
+                                        theta, tau_bq, tau_payroll, h_wealth,
+                                        p_wealth, m_wealth, retire, T, S, J)
+        e           = [T,S,J] array, effective labor units
+        lambdas     = [J,] vector, population weights by lifetime income group
+        omega       = [T,S] array, population weights by age
+        method      = string, 'SS' or 'TPI'
+        etr_params  = [T,S,J] array, effective tax rate function parameters
+        theta       = [J,] vector, replacement rate values by lifetime income group
+        tau_bq      = scalar, bequest tax rate
+        h_wealth    = scalar, wealth tax function parameter
+        p_wealth    = scalar, wealth tax function parameter
+        m_wealth    = scalar, wealth tax function parameter
+        tau_payroll = scalar, payroll tax rate
+        retire      = integer, retirement age
+        T           = integer, number of periods in transition path
+        S           = integer, number of age groups
+        J           = integer, number of lifetime income groups
+    Functions called:
+        tau_income
+        tau_wealth
+    Objects in function:
+        I    = [T,S,J] array, total income
+        T_I  = [T,S,J] array, total income taxes
+        T_P  = [T,S,J] array, total payroll taxes
+        T_W  = [T,S,J] array, total wealth taxes
+        T_BQ = [T,S,J] array, total bequest taxes
+        T_H  = [T,] vector, lump sum transfer amount(s)
+    Returns: T_H
+
+    '''
+
+    e, lambdas, omega, method, etr_params, theta, tau_bq, \
+        tau_payroll, h_wealth, p_wealth, m_wealth, retire, T, S, J,\
+        tau_b, delta_tau = params
+
+    I = r * b + w * e * n
+
+    if I.ndim == 2:
+        T_I = np.zeros((S,J))
+        for j in xrange(J):
+            TI_params = (e[:,j], etr_params)
+            T_I[:,j] = tau_income(r, w, b[:,j], n[:,j], factor, TI_params) * I[:,j]
+    if I.ndim == 3:
+        T_I = np.zeros((T,S,J))
+        for j in xrange(J):
+            if etr_params.ndim == 3:
+                tau_inc_params3D = etr_params[:,j,:]
+            if etr_params.ndim == 4:
+                tau_inc_params3D = etr_params[:,:,j,:]
+            TI_params = (e[:,:,j], tau_inc_params3D)
+            T_I[:,:,j] = tau_income(r[:,:,j], w[:,:,j], b[:,:,j], n[:,:,j], factor, TI_params) * I[:,:,j]
+    T_P = tau_payroll * w * e * n
+    TW_params = (h_wealth, p_wealth, m_wealth)
+    T_W = tau_wealth(b, TW_params) * b
+    if method == 'SS':
+        T_P[retire:] -= theta * w
+        T_BQ = tau_bq * BQ / lambdas
+        biz_params = (tau_b, delta_tau)
+        business_revenue = get_biz_tax(w, Y, L, K, biz_params)
+        REVENUE = (omega * lambdas * (T_I + T_P + T_BQ + T_W)).sum() + business_revenue
+    elif method == 'TPI':
+        T_P[:, retire:, :] -= theta.reshape(1, 1, J) * w[:,retire:,:]
+        T_BQ = tau_bq.reshape(1, 1, J) * BQ / lambdas
+        biz_params = (tau_b, delta_tau)
+        business_revenue = get_biz_tax(w[:T,0,0], Y, L, K, biz_params)
+        REVENUE = (omega * lambdas * (T_I + T_P + T_BQ + T_W)).sum(1).sum(1) + business_revenue
+    return REVENUE
